@@ -1,2 +1,3 @@
-"""Playwright MCP Server - Python Implementation"""
+"""Playwright MCP Server - Python Implementation."""
+
 __version__ = "1.0.0"
